added mesoscope mode with arduino
added a new mode of operation called mesoscope_mode, set in djPrefs or using a button on the  front panel.  We can now choose to run in classic OpenEphys mode which requires an intan board and communicates with OpenEphys to start acquisition and get a data directory, or the new mesoscope mode which connects to an arduino over a usb port, and delivers a protocol start TTL when you click record (to trigger the mesoscope and spike2), and a soundcard trigger for each sound. notebook and stimlog are stored locally in a data directory. In mesoscope mode there is no zmq communication with open ephys.